Move imaging table to cpp (cherry picked from commit a17bc4f)
